requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
boto3==1.26.13
botocore==1.29.13
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1.3
django-environ==0.9.0
django-request-logging==0.7.5
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.32
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
redis==4.3.4
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
uuid==1.30
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zxcvbn==4.4.28
